// src/payload-types.ts
export interface Tenant {
  id: string;
  /**
   * The name of the store (e.g., John Doe's Store)
   */
  name: string;
  /**
   * The subdomain of the store (e.g., [slug].funroad.com)
   */
  slug: string;
  image?: (string | null) | Media;
  /**
   * Stripe Account ID associated with your shop
   */
  stripeAccountId: string;
  /**
   * You cannot create products until you’ve submitted your Stripe details.
   */
  stripeDetailsSubmitted?: boolean | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  tenant?: (string | null) | Tenant;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}